Fix for shared NIC with the N3000 FPGA
An issue can occur when using an N3000 FPGA NIC with upper
vlan interfaces (oam, mgmt, etc).  Because the N3000 is reset
after the networking initialization is complete, any IPv6
addresses, or (IPv4 / IPv6) default routes configured on the
vlan interfaces will be removed.

This commit fixes the issue by operating on a generated
FPGA resource containing the N3000 interface name and any
of the 'used by' (upper) interfaces belonging to the device.

If these resources are present, puppet will restart the
upper interfaces after the N3000 FPGA is reset, allowing
any configured IP addresses and routes to be restored.

Closes-Bug: #1931461
Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I2518255f7c33f1914fa5682ca8e00afb2582d8d7