Add a puppet rehome manifest
This commit adds a subcloud_rehome manifest, this manifest is to
re-configure the keystone authentication passwords of related services
and reload these services.

Tested with apply this manifest after the keystone authentication
passwords migrated in secure static hiera data. The passwords in
services' configuration files are modified and the related services
are reloaded after the apply.

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/795111
Story: 2008774
Task: 42581
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: I93902da0939f5b6857e58f9fd6ada6a9d5c1a3ab