Introduce runtime extension classes for patchability
This update introduces a set of empty extension classes that can be
used in patches to safely introduce new runtime classes. The patch can
update an unused extension class to include the new runtime class,
allowing sysinv to be patched to apply the selected extension class
rather than the new runtime class directly. This allows sysinv to
safely apply the class on a remote host, even if the host has not yet
been patched.

Closes-Bug: 1931270
Signed-off-by: Don Penney <don.penney@windriver.com>
Change-Id: I302d7100a52d085aa5c58f80173edb5cfd490054