Avoid SRIOV misconfiguration at FEC device
When performing a batch SRIOV configuration (e.g. from shell script),
it may be possible that commands consecutively issued - either related
to interface(s) or a second FEC device settings - make the FEC device
configuration to be not fully applied, and then unlock fails by
semantic check.

The reason for that is hieradata might get updated by intermediate
inventory reports that could overwrite the values to be set with the
current ones, since it hasn't separate fields for both values. This
change makes 'num_vfs' to be used as the "configured" value (like it is
actually used in puppet manifest) and creates another 'sriov_vfs' that
will contain the current value for number of VFs.

Another change is not calling platform::devices::fpga::fec::vf when
applying runtime manifest, which could also fail in case PCI addresses
are changing from what is in hieradata.

Depends-On: https://review.opendev.org/c/starlingx/config/+/795850
Closes-Bug: #1931458
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I97c4751f9f977468d26bb2e6fea62ed7a6fa75cb