Reset N3000 device based solely on its presence
Instead of resetting N3000 device in the AIO manifest when the PCI
address of some is(are) reported back by sysinv-[fpga-]agent, this
change introduces a new facter to look for the device at system
and flag its existence. Then, during AIO manifest apply, that flag
is checked and N3000 is reset (if not yet done) regardless of any
configuration already available at database/hieradata.

Partial-bug: #1933119
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: Ibdb2b1a77b9ec3a3a516e5cf28da59e9ac74be60