Add proxyconfig class for containerd
This change adds a proxyconfig class for containerd. Containerd keeps
its own proxy config. This proxy config needs to be changed in
addition to the docker proxy config when we change the proxy config
as a whole. The containerd proxyconfig class is used by the docker
runtime class when changing the proxy config without the need for
a node lock unlock.

Change-Id: I7b08f04ffa7b20beffa77e02188d6df3fe7809b9
Partial-Bug: 1931593
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>