Update the kubernetes bind mounts based on the desired versions
Before calling kubeadm/kubelet/kubectl we want to ensure that the
bind mounts are properly set up so that we get the version of the
command that is appropriate.

We probably don't actually need to check the bind mounts in
class platform::kubernetes::kubeadm since on initial install the
ansible playbooks will set it, and during a kubernetes upgrade
the runtime manifests should set them and persist the changes to
/etc/fstab.  That said, it won't hurt anything and won't add much
delay to double-check on host startup.

We also may not need to check the bind mounts in
class platform::kubernetes::upgrade_first_control_plane since the
kubeadm mount for the first controller should have been updated in
class platform::kubernetes::pre_pull_control_plane_images and the
kubelet mount doesn't get updated until
class platform::kubernetes::master::upgrade_kubelet.  But again
it shouldn't cause problems to double-check it.

Story: 2008972
Task: 42880
Change-Id: I3ec63c7ed069e521df320ebddb2e71aa8a9eedfe
Depends-On: https://review.opendev.org/c/starlingx/config/+/802302
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>