check value of kubernetes host labels
Before this commit the logic was just based on
checking that a key exists in the hieradata
(e.g. openstack-compute-node) but its value was
not available there, so the value can not be
checked to see if it is enabled or disabled.

Another commit added support for the values of
the labels in the hieradata and this commmit
addresses the changes required in the puppet
code to make checks based on both key and
its value.

Closes-bug: 1940042
Change-Id: Iefbcb1cde7db27250d9a3629472810d20f202048
Signed-off-by: Daniel Safta <daniel.safta@windriver.com>
Depends-on: https://review.opendev.org/c/starlingx/config/+/804526