FEC devices enabled by default in puppet manifest
In the context of making pf-bb-config tool being launched to configure
also N3000 devices (https://storyboard.openstack.org/#!/story/2009036)
the puppet manifest relies on the 'enabled' parameter which is set via
a sysinv script based on PCI devices inventory. It was observed that if
this feature is available not by a fresh install but instead through a
patch applied on the corresponding modules, it required an extra reboot
to have the hieradata populated with that new parameter, which is unde-
sired in installed systems.

This change makes the puppet manifest to consider the FEC devices are
enabled by default. By doing that, the first run of the manifest after
applying the patch won't rely on the (absent) hieradata parameter, but
will still trigger the launch of pf-bb-config tool to configure the FEC
device. Later on, it is expected that 'enabled' parameter is added to
hieradata and manifest can rely on that for any further configuration.

NOTE: actually, this change makes puppet manifest to match the same
behaviour of the corresponding at database. There, 'enabled' field for
devices is 'True' by default.

Task: 43030
Story: 2009036
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I65d1c6cc34029b20a4777dbdafa0cddf04082206