Add error parsing on dcorch WSGI pipeline
Since dcorch.api.proxy does not use pecan hooks approach, it is
required to keep the current architecture and add an additional
middleware in the dcorch WSGI pipeline.
It was added a "parser" in the beginning of the pipeline to parse
errors on the pipeline and add the proper "faultstring" field to
be parsed by the cgts-client.

Test Plan: --os-region-name SystemController load-import
PASS: Verify proper error when no space available
PASS: Verify proper load-import when enough space available

Story: 2009285
Task: 43576
Closes-Bug: 1945737
Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/813269
Change-Id: I35551ecb82993f607d26949f7bf838fa07175245
Signed-off-by: Adriano Oliveira <adriano.oliveira@windriver.com>