Wait kube-apiserver on kube-rootca-update trustcas
During both trust-cas phases (trust-both-cas and trust-new-ca) of the
kube-rootca-update, the kube-apiserver is restarted in order to trust
the new rootcas in place, but it doesn't currently wait for the k8s API
server to be properly restarted before completing the command.
This way, running the subsequent command, kube-rootca-pods-update,
without giving the right amount of time for the kube-apiserver to be
properly restarted causes it to fail and get stalled.

Adding a wait condition at the end of both trust-both-cas and
trust-new-ca fixes this behavior by not allowing the next steps to be
run until the kube-apiserver is ready.

The check used is as suggested in in the k8s health checks doc:
https://kubernetes.io/docs/reference/using-api/health-checks/

Test Plan:

PASS: Verify kube-apiserver is reachable immediately after
kube-rootca-host-update trust-both-cas and trust-new-ca completes
PASS: Verify kube-rootca-pods-update behaves correctly running
immediately after host-rootca-host-update completes
PASS: Verify kube-rootca-host-update step fails if kube-apiserver fail
to get online before the timeout (maximum of 180 s)

Closes-Bug: 1950004
Signed-off-by: Rafael Camargos <RafaelLucas.Camargos@windriver.com>
Change-Id: Ia15f5e9f46311203f9c4f89ce6b7d18b04356886