Wrap IPv6 in square brackets in ceph.conf
New version of ceph requires IPv6 to be wrapped in square brackets

When running a B&R procedure on ipv6 setups, the restore playbook
fails.
This happens because the restore playbook has to add the v1 protocol
port number in the public_addr option for the monitor group in order
to start the ceph monitor. It fails in ipv6 networks because ipv6
addresses must be wrapped in square brackets and they are not.

How it is (example):
  [mon.controller-0]
  public_addr = fd01::2

How it should be (example):
  [mon.controller-0]
  public_addr = [fd01::2]

Test Plan:
 - AIO-SX ipv4 B&R
 - AIO-DX ipv4 B&R
 - AIO-SX ipv6 B&R

Story: 2009074
Task: 44044
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>
Change-Id: Ia16a5b98e5b880654f7f9395b4c904dab9314cc0