Wait for kube apiserver after apply service parameter
When doing a: "system service-parameter-apply kubernetes"
the kube apiserver will be restarted but the service
parameter command will return regardless of the apiserver
state. And if we run some 'kubectl' commands after that, they will
fail until the apiserver is back online.
To avoid these bad scenarios, we will wait for the kube apiserver
to be up after we apply the service parameter.

Closes-Bug: 1953183
Change-Id: Ic3cc3728a8eb363a06f268c904680f30c40b5d44
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>