Correct ordering when upgrading etcd certs
During an upgrade activate we generate new etcd certs. As part of this
process we need to restart etcd and direct the kube-apiserver to use the
new certs.

With this update we'll restart the etcd server before updating the
kube-apiserver. Finally we'll wait for the kube-apiserver to be ready.
Waiting for the apiserver is necessary so subsequent sysinv operations
can interact with kubeadm.
https://bugs.launchpad.net/starlingx/+bug/1954488 and
https://bugs.launchpad.net/starlingx/+bug/1953183 are seen if we do not
wait for the apiserver.

Note curl -f is used to ensure the readyz call returns 200/OK.

Closes-Bug: 1954488
Closes-Bug: 1953183
Change-Id: I3b31153d8779262c0bf5e57526b41421689bad36
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
(cherry picked from commit 61b72946b8eb81a34c7559b4a0015a3af23f179e)