Fix for SR-IOV VF rate limiting on N3000 FPGA NICs
It is possible that a user may encounter puppet failures on system
lock/unlock after configuring tx max rate-limiting on an SR-IOV
interface belonging to an N3000 FPGA.

The problem may occur if puppet configures the max tx rate on the
VF(s) after the device has been reset, but before it has been
brought down/up (which causes the VFs to be created).

This commit fixes the issue by adding an explicit dependency
that ensures the rate limit setting for each VF occurs after
it's driver binding (this is the more appropriate thing to
do even outside of the context of the N3000 issue).

Since there is an existing dependency on the N3000 for the
driver bindings to occur after the device reset/ifdown/ifup,
the particular issue noted above is resolved.

Testing:

PASS: examine the puppet logs and ensure the relationship
      between the driver binding and rate-limit setting for each
      configured VF has been noted.
PASS: ensure the system unlocks successfully and that the
      VFs have the appropriate tx max rate setting and have
      been associated with the appropriate driver.
PASS: tested the rate-limit setting with VF drivers vfio and
      netdevice (kernel driver)

Regression:

PASS: ensure that if a user has not configured rate-limiting
      that the dependency is noted, but the rate-limiting
      action is not performed.
PASS: ensure that the system behaves as before for SR-IOV
      interfaces that are not attached to an N3000 FPGA.

Closes-Bug: 1956604
Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I5321aa94c8bc0e93fcc107429d44f294da1d17ff