Enable unsafe_noiommu_mode if vswitch enabled
A failure in the stx-puppet manifest was introduced in [1]. The
failure occurs because the vfio-pci driver is not loaded when vswitch
is disabled and it is a required kernel module for the
unsafe_noiommu_mode exec "echo" command.

Test plan:
PASS: Verify that in a hybrid cluster environment (openstack nodes and
k8s workers) after removing the vswitch host label the the vswitch is
successfully disabled.

[1] https://review.opendev.org/c/starlingx/stx-puppet/+/818567
Story: 2009702
Task: 44289
Signed-off-by: Iago Estrela <IagoFilipe.EstrelaBarros@windriver.com>
Change-Id: I480b99a1cd09cb3db44bed0f30b09fef0036c801