[PTP dual NIC config] Enable ptpinstance manifest
This commit makes changes to the ptpinstance manifest to support
runtime application as well as enabling the manifest for controllers,
workers, storage and aio configs.

This does not disable the single instance ptp functionality, so as not
to break exiting deployment manager configs. A future review will handle
removing the single instance ptp manifests and deprecating the related
CLI/APIs.

Test Plan:

Pass: Build and install
Pass: Verify ptpinstance functionality works and can be applied
as a runtime manifest.

Regression:

Pass: ensure that single instance ptp is unaffected.

Story: 2009248
Task: 44293

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I064d31f87279a3c290f7ff3d76895546af15bf6a