Fix resource lookup for ovs
On change ae635b5b80fcb61c429a6fc17961a9f3bf614964, the vswitch_class
was changed to ovs_dpdk, but the resources created by sysinv at [1]
are at the platform::vswitch::ovs:: lookup. This mismatch is failing
lookup and the bridges for the underlying datanetworks aren't being
created when puppet runs. As a result, the neutron-ovs-agent pods are
failing with CrashLoopBackoff. This commit fixes it by reverting the
resources to the correct lookup path on hiera.

[1] https://opendev.org/starlingx/config/src/commit/ece13f740847f3bcc7470cc7ec8c1896dd61f014/sysinv/sysinv/sysinv/sysinv/puppet/ovs.py#L108

TEST PLAN
PASS ovs-dpdk: Clean install of the Starlingx ISO verified that the
     br-phy* bridges were created for the underlying datanetworks
     using ovs-vsctl on the host
PASS ovs-dpdk: Installation of stx-openstack is successful
PASS ovs-dpdk: Created project networks and instances
PASS ovs: Clean install of the Starlingx ISO
PASS ovs: Installation of stx-openstack is successful and the
     openvswitchd and ovs-agent pods are running OK
PASS ovs: Created project networks and instances

Closes-Bug: #1958073
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I53e16df5403fa7c7f82b8e67e3e5d18a2103d599