[PTP SyncE] Apply synce command to configured port
This commit fixes an issue where synce recovered clock commands were
being applied to the base port of the NIC rather than the specified
port. On Westport Channel NICs, many of the configuration commands must
be directed at the base port. The rclka and rclkb commands are the
exception and should instead be applied directly to the specified port.

This change also includes an update to the format of the clock-conf.conf
file which now indicates both the configured port and the base port for
each set of supplied parameters. This allows collectd to correctly parse
which ports should be checked.

The node manifests have been updated to include the nic_clock task.

Added a notify parameter to the task that generates the environment
variables file in order to trigger a service restart when cmdline_opts
is added or modified.

The final change was to relocate the task to create clock-conf.conf to
outside of the create_resources statement as there is no need for it to
be run multiple times.

Testing:

Full build and install, validated NIC configuration could be applied,
ptp behaviour unaffected.

Story: 2009130
Task: 44475
Task: 44508
Task: 44509

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I1e8f3ed656c6ea2d5100988d096a7f645591d647