Change eventlet.wsgi.server loglevel for dcmanager
The level of eventlet.wsgi.server logs was changed to WARN to avoid
printing the the default request logs. The request logs will be printed
by specific hook of the server. Note: this change will be set as
dependant of the change that creates the hook that prints the request
logs, in order to disable the default logs when the new ones are
enabled.

Test Plan:

PASS: Successfully deploy a Distributed Cloud and verify that the logs
of DC Manager API requests no longer appear in DC Manager log files.

Story: 2009824
Task: 44549

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/829396
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I49ae9286f71620b2b3264b0b58b5e99eb5cae29f