Create mon runtime puppet class
Create the class platform::ceph::mon::runtime to be used to change at
runtime if ceph will accept insecure global id reclaim.

Using this a caller can change the param $auth_id_reclaim (which
defaults to false). When the value is true id reclaim is allowed and
the respective alarms are disabled, otherwise, if the param is false id
reclaim is forbidden and the respective alarms are enabled.

Test plan:
Apply Sysinv service-parameter which changes param $auth_id_reclaim in
three states:
    - True
    - False
    - Null

Story: 2009867
Task: 44562
Change-Id: I856924c56c2a6005d9e1889d2a6b0fb049b00b08
Signed-off-by: Delfino Curado <delfinogomes.curadofilho@windriver.com>