Remove usage of forward-journald
This is a logging tool that was created as a workaround
to a bug where docker would crash when journald was stopped
(https://bugzilla.redhat.com/show_bug.cgi?id=1300076).

The bug has since been fixed on docker and tests have shown
etcd does not present this bug at the moment,
so forward-journald can be removed.

Test Plan:

Centos
Pass: etcd service started normally
Pass: Logs saved to /var/log/kern.log as before change
Pass: etcd service does not crash when systemd-journald is stopped

Debian
Pass: etcd service started normally
Pass: etcd service does not crash when systemd-journald is stopped

Note:
Some of the etcd logs that show up on centos at /var/log/kern.log
are not showing up for debian.
Tested adding forward-journald to debian but the behavior
is the same with or without it.

Story: 2009101
Task: 44603

Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>
Change-Id: I7ea13b4dde212322aab31c9e0469f87c8762e7d8