Extend cert-mon network_max_retry default value
We are running into cases where slow hardware means that subclouds can
take a long time to complete their startup. In such cases, cert-mon
is giving up on the initial audit after 15m, leaving the subcloud in
an out-of-sync state until the next daily cert-mon audit run.

The network_max_retry works with the network_retry_interval (in seconds)
to affect the total time spent retrying a subcloud before giving up:

  network_retry_interval * (network_retry_interval/60) = time in minutes

So we are increasing the overall retry time from 15m to 90m:

   30 retries * 180s/60m per retry = 90m

Test Plan:

PASS:
- install using new values, verify that the default is now changed
- verify retry limit is reached
- verify case where daily audit is initiated while the subloud is in the
  reattempt state

Closes-Bug: 1964811

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I3d279a4c0f9f041047df4b5a3e850396d91bbb24