Fix vswitch manifest enabled parameter default value
Since vswitch is not a required service for the platform
we can disable it by default. This prevents the dependency
of the manifest on the OVS plugin hieradata, which may provide
outdated data during patch apply.

Test plan:
PASS: Host lock and unlock with enabled default to undef.
PASS: Bootstrap with enabled default to undef.
PASS: Bootstrap with vswitch_type=none.
PASS: Bootstrap with vswitch_type=ovs-dpdk.

Closes-Bug: 1965143

Signed-off-by: Iago Estrela <IagoFilipe.EstrelaBarros@windriver.com>
Change-Id: Ia07877d2188b681dfcf9c2700222c1cc7f73f7e6