[PTP SyncE] Disable NTP on controller when PTP is enabled
The ntp puppet manifest was incorrectly starting NTP on system
controllers when PTP is enabled, which results in two services trying to
adjust clocks and causes instability in the PTP time. This change makes
it so that NTP is not enabled when the host clock_synchronization
parameter is set to ptp.

Tests:

PASS: Configured NTP on host, lock/unlock, confirm ntp is running, PTP
is not.
Switch to PTP, lock/unlock, confirm PTP is running and NTP is not.

Story: 2009130
Task: 44814

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I64262843ec8bb83d1b6e4ce1dfa19eafdcbc098f