Exclude bnxt_en and bnxt_re from kdump initramfs
This commit ensures that the bnxt_en and bnxt_re drivers are excluded
from the kdump initramfs, because the kdump kernel's memory is limited,
and there is a chance that including bnxt_en/bnxt_re drivers could
re-trigger an issue similar to StarlingX bug 1957938 on Launchpad, where
vmcore dumps did not work due to the kdump kernel running out of memory.

Based on a review of dracut's code, the "hostonly" mode is enabled by
default, and in hostonly mode all wired Ethernet adapter drivers loaded
at the time of dracut's execution are considered for inclusion in the
initramfs.

To be safe, exclude these drivers from the kdump environment.

Testing:
- An ISO image is built using a monolithic incremental build procedure.
- The ISO image is installed into a VM in low-latency All-in-One Simplex
  mode, and the installation is bootstrapped.
- After bootstrap, /etc/kdump.conf includes the directive to omit the
  bnxt_en and bnxt_re drivers as well as the drivers that used to be
  omitted.

Story: 2009915
Task: 44836

Depends-On: https://review.opendev.org/c/starlingx/kernel/+/833464

Change-Id: Ic600614d08614a5768195665ea19ffc8a6be02a9
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>