Fix route configuration in DC setup
In distributed cloud (DC) systems it was observed that configuration of
routes used to connect the DC controller(s) to the subclouds was failing
since the proper function 'update_routes' was not been found in the
proper child script (either for CentOS or Debian).

This change fix it by performing the same verification to detect what
script should be sourced, depending on the base distro being used.

Test Plan:
PASS: Install in DC enviroment (dc-controller: 2 controllers+1 compute,
      plus 2 AIO-SX subclouds) and checked connectivity via DC
      management routes.

Task: 44840
Story: 2009101
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I21524b75c401a6d9de03e1efe71555f54f62b9a3