Use realpath to check the existing mounts
This work is part of the Debian integration effort.
This will remove a bootstrap issue on Debian.

On Debian /mnt and /opt are symlinks. On CentOS realpaths.
Mount will create the mountpoint at the realpath (both OSs).
On Debian this is an issue because the mountpoints are now on a
symlinked path, creating a false negative. Due to the false
negative puppet will run code that is not supposed to run, messing
the logic downstream.

CentOS tests:
PASS: AIO-SX bootstrap
PASS: AIO-SX unlock
Also checked that mount entrypoints are already realpath on an
existing AIO-DX after unlock, and on an existing AIO-SX before
bootstrap, after bootstrap and after unlock.

Debian tests:
PASS: AIO-SX bootstrap
PASS: AIO-SX unlock

Story: 2009101
Task: 44913
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: Ib9597fd3231f380adc3373fad18d0d17ba088deb