Debian: Avoid setting up yum repos for patching
Yum repo initialization in puppet for patching can be
bypassed on Debian.  The Debian patching solution is based
on ostree.

The changes in puppet to support patching on Debian through
ostree will merge when that feature is implemented.

Test Plan:
   Verify that Centos environment can be boostrapped
amd unlocked and the yum file is created.
   Verify that Debian environment can be bootstrapped
and unlocked.

Story: 2009101
Task: 44936
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ibde05d37e80e1ef2055a2512963b0484a1e26fe6