Add flag to ensure PF is up before SR-IOV set up
As part of the integration of the bnxt_en driver, it was discovered that
it requires a PF to be administratively up before any VFs can be
configured. Currently the system tries to write to the VF file straight
away, without checking the interface state beforehand.

This change modifies the puppet that enables SR-IOV, adding a config
flag to it that when enabled will try to to ensure a port is
administratively up before trying to write to the VF file.

TESTS:
The tests below were run on a host with a BCM57504 NIC
- Installation is successful
- VFs can be created successfully through system commands
- VFs do not throw any errors when set up
- VFs are set up correctly after a lock/unlock

Depends-On: https://review.opendev.org/c/starlingx/config/+/836958

Story: 2009915
Task: 45013
Signed-off-by: Antonio Augusto Vilas Boas Teixeira
<antonio.augustovilasboasteixeira@windriver.com>
Change-Id: Iea001ea04fc199043e941b4c65b6c84f6a679451