Add support for bonded interfaces on Debian
Sysinv is generating the stanza "allow-[master] [slave]" as part of
the options of network_config. This is necessary because it does not
have support on the puppet-network module. Since this script already
does the parsing to generate the definitive files, an extra loop is
executed to identify files with the stanza and move it to the end.
This is necessary to not break the parsing by ifup.

It was also detected that get_prefix_length was not calculating IPv6
prefix length, only IPv4. This change corrects this

Test Plan:
PASS configuration of balanced interfaces with all txhashpolicy
     values
PASS configuration of active-standby interfaces with all
     primary-reselect values
PASS configuration of addresses and routes on the bond interface
PASS configuration of VLAN interface on top of bond interface, with
     static addresses and routes
PASS all tests executed with host lock/unlock cycle

Story: 2009965
Task: 45197
Depends-On: https://review.opendev.org/c/starlingx/config/+/839619

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I5895c371e6e5c63ad3ed89b2551bc1d809a37438