[PTP SyncE] Set ice-gnss thread prio in puppet
Problem: The ice-gnss thread is not spawned until ptp services are
configured, so on first time setup, the affine-process.sh script is not
able to set the priority of the thread because it runs before puppet.

This change adds a task to the puppet manifest to also set the niceness
of the ice-gnss threads in order to handle the case of first time setup.
Subsequent lock/unlocks will handled earlier in the startup process by
affine-process.sh.

See https://review.opendev.org/c/starlingx/utilities/+/839795 for the
earlier change to affine-process.sh.

Testing:

Pass: Thread niceness set correctly on first time setup on AIO-SX.
Niceness is also correctly set on subsequent lock/unlocks. Switching
node from ptp to ntp and back also results in correct priority.

Closes-Bug: 1970776

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I1c9c0ffb6cd0dad7b77232522832b1645256dcfd