debian: Use pg_createcluster to set up postgresql
Postgresql behaves differently on Debian compared to
Centos 7, it uses a pg_createcluster script to wrap
the postgresql commands such as initdb in order
to manage the postgresql database.

In Centos 7 we just use the initdb command run as the postgres
user in order to create the database.

Here we detect if we are using either Centos or Debian and do
the appropriate setup for the respective distro.

Also in this commit we enable IPV6 for postgresql as well.

Test Plan
PASS Build puppet-manifests package
PASS Build ISO
PASS Unlock system
PASS Build Centos ISO
PASS bootstrap Centos 7

Story: 2009101
Task: 43317

Depends-On: https://review.opendev.org/c/starlingx/utilities/+/840497

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Iba0ef50610e703b3caf044024a21609a1a713096