Do not abort the network script if the loopback is unconfigured.
This protection was added on change:
https://review.opendev.org/c/starlingx/stx-puppet/+/752081

But it is deemed unnecessary as it is possible to set the loopback to
class none. On bootstrap, the loopback is set to the class platform
and received by the mgmt cluster-host networks. This initial setup
can then be modified, in duplex installations, by moving the networks
to physical interfaces and leaving the loopback as class none (but
the operator can leave it as class=platform).

With the loopback class set to none, the puppet-network module will
not generate the file /var/run/network-scripts.puppet/ifcfg-lo, which
will result in the AIO manifest applying error.

There are already several validations on unlock to prevent invalid
configurations (e.g., one cannot unlock if the mgmt network is not
anchored on a platform interface). To add this validation to this
internal script is just preventing the execution without actually
protecting the system.

On the tests below the management and cluster-host network were moved
to the physical interfaces.

Test Plan:
PASS  Set interface loopback to class none and unlock

No side effects were detected.

Closes-bug: 1973614


Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I17f4070667e2dabca2cedc059cb0609cee444092