debian: temporary disable multipath by default
It is observed during a lock-unlock that multipath fails to start on
Debian.

Disable multipath by default to allow Debian testing.
There will be a follow-up commit to fix the issue and re-renable
multipath by default.

Tests on AIO-SX Debian:
PASS: unlock
PASS: lock-unlock

SKIP CentOS tests because it is just a refactor.

Story: 2010046
Task: 45431
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I034cf2747d3f12214c5008a5b6a74abb13dadd03