Debian: only update routes after interfaces are configured
In Debian it was detected a dependency to create interfaces before
the creation of routes. This change sets this order for the Debian
installations, and keeps the current order for CentOS.

Detected that is all static routes are removed the file
/var/run/network-scripts/routes is not generated, we need to
remove all route in /etc/network/routes. This was not happening

Also adding log messages if the route creation/delete fails.

Test Plan
PASS  add routes with system host-route-add and unlock on Debian
PASS  del routes with system host-route-delete and unlock on Debian

Closes-Bug: 1977558

Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Change-Id: I72714349957c6a2241485907ea79f8b3e6fe6537